+++ src/settings/MappingProfiles/detailsSections/edit/OrderDetailSection/Vendor.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import PropTypes from 'prop-types';
 import {
   FormattedMessage,
@@ -41,6 +41,14 @@ export const Vendor = ({
 }) => {
   const { formatMessage } = useIntl();
 
+  const VENDOR_FIELDS_MAP = {
+    VENDOR_REF_NUMBERS: 44,
+    VENDOR_REF_NUMBER: index => getSubfieldName(VENDOR_FIELDS_MAP.VENDOR_REF_NUMBERS, 0, index),
+    VENDOR_REF_TYPE: index => getSubfieldName(VENDOR_FIELDS_MAP.VENDOR_REF_NUMBERS, 1, index),
+    ACCOUNT_NUMBER: 45,
+    INSTRUCTION_TO_VENDOR: getFieldName(46),
+  };
+
   const vendorRefTypeOptions = [
     {
       label: formatMessage({ id: `${TRANSLATION_ID_PREFIX}.order.vendor.field.continuationRefNumber` }),
@@ -64,7 +72,15 @@ export const Vendor = ({
     value: accountNo,
   }));
 
-  const vendorDetailFieldIndex = 47;
+  const handleVendorNumbersAdd = useCallback(
+    () => onAdd(vendorRefNumbers, 'vendorDetail', VENDOR_FIELDS_MAP.VENDOR_REF_NUMBERS, initialFields, setReferenceTables, 'order'),
+    [VENDOR_FIELDS_MAP.VENDOR_REF_NUMBERS, initialFields, setReferenceTables, vendorRefNumbers],
+  );
+
+  const handleVendorNumbersClean = useCallback(
+    index => onRemove(index, vendorRefNumbers, VENDOR_FIELDS_MAP.VENDOR_REF_NUMBERS, setReferenceTables, 'order'),
+    [VENDOR_FIELDS_MAP.VENDOR_REF_NUMBERS, setReferenceTables, vendorRefNumbers],
+  );
 
   return (
     <Accordion
@@ -74,8 +90,8 @@ export const Vendor = ({
       <RepeatableField
         fields={vendorRefNumbers}
         addLabel={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.vendor.field.vendorRefNumbers.addLabel`} />}
-        onAdd={() => onAdd(vendorRefNumbers, 'vendorDetail', vendorDetailFieldIndex, initialFields, setReferenceTables, 'order')}
-        onRemove={index => onRemove(index, vendorRefNumbers, vendorDetailFieldIndex, setReferenceTables, 'order')}
+        onAdd={handleVendorNumbersAdd}
+        onRemove={handleVendorNumbersClean}
         renderField={(field, index) => (
           <Row left="xs">
             <Col xs={6}>
@@ -84,7 +100,7 @@ export const Vendor = ({
                   <Field
                     component={TextField}
                     label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.vendor.field.vendorRefNumber`} />}
-                    name={getSubfieldName(vendorDetailFieldIndex, 0, index)}
+                    name={VENDOR_FIELDS_MAP.VENDOR_REF_NUMBER(index)}
                     validate={[validation]}
                   />
                 )}
@@ -93,7 +109,7 @@ export const Vendor = ({
             <Col xs={6}>
               <AcceptedValuesField
                 component={TextField}
-                name={getSubfieldName(vendorDetailFieldIndex, 1, index)}
+                name={VENDOR_FIELDS_MAP.VENDOR_REF_TYPE(index)}
                 label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.vendor.field.vendorRefType`} />}
                 optionValue="name"
                 optionLabel="label"
@@ -109,13 +125,13 @@ export const Vendor = ({
         <Col xs={6}>
           <AcceptedValuesField
             component={TextField}
-            name={getFieldName(48)}
+            name={getFieldName(VENDOR_FIELDS_MAP.ACCOUNT_NUMBER)}
             label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.vendor.field.accountNumber`} />}
             optionValue="value"
             optionLabel="label"
             wrapperLabel={`${TRANSLATION_ID_PREFIX}.wrapper.acceptedValues`}
             setAcceptedValues={setReferenceTables}
-            acceptedValuesPath={getAcceptedValuesPath(48)}
+            acceptedValuesPath={getAcceptedValuesPath(VENDOR_FIELDS_MAP.ACCOUNT_NUMBER)}
             acceptedValuesList={accountNumbersOptions}
             hasLoaded
           />
@@ -126,7 +142,7 @@ export const Vendor = ({
               <Field
                 component={TextArea}
                 label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.vendor.field.instructionsToVendor`} />}
-                name={getFieldName(49)}
+                name={VENDOR_FIELDS_MAP.INSTRUCTION_TO_VENDOR}
                 validate={[validation]}
               />
             )}